♻️ refactor(api): simplify entity import path
