+++ apps/api/src/application/use-cases/auth/interfaces/signup.usecase.interface.ts
@@ -1,6 +1,5 @@
 import type { Err, Result } from "../../../../common/utils";
-import type { Session } from "../../../../domain/entities/session";
-import type { User } from "../../../../domain/entities/user";
+import type { Session, User } from "../../../../domain/entities";
 
 export type SignupUseCaseSuccessResult = {
 	user: User;
