+++ apps/api/src/application/use-cases/email-verification/email-verification-confirm.usecase.ts
@@ -1,5 +1,5 @@
 import { err } from "../../../common/utils";
-import type { User } from "../../../domain/entities/user";
+import type { User } from "../../../domain/entities";
 import type { IEmailVerificationCodeRepository } from "../../../interface-adapter/repositories/email-verification-code";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
 import type {
