+++ apps/api/src/interface-adapter/gateway/oauth-provider/oauth-provider.gateway.ts
@@ -1,4 +1,4 @@
-import type { OAuthProvider } from "../../../domain/entities/oauth-account";
+import type { OAuthProvider } from "../../../domain/entities";
 import type { OAuthProviderEnv } from "../../../modules/env";
 import type { IOAuthProviderGateway } from "./interfaces/oauth-provider.gateway.interface";
 import { DiscordOAuthGateway } from "./providers/discord.gateway";
