+++ apps/api/src/interface-adapter/repositories/email-verification-code/email-verification-code.repository.ts
@@ -1,5 +1,5 @@
 import { and, eq, lte } from "drizzle-orm";
-import { EmailVerificationCode } from "../../../domain/entities/email-verification-code";
+import { EmailVerificationCode } from "../../../domain/entities";
 import type { DrizzleService } from "../../../infrastructure/drizzle";
 import type { IEmailVerificationCodeRepository } from "./interfaces/email-verification-code.repository.interface";
