+++ apps/api/src/interface-adapter/repositories/email-verification-code/interfaces/email-verification-code.repository.interface.ts
@@ -1,4 +1,4 @@
-import type { EmailVerificationCode } from "../../../../domain/entities/email-verification-code";
+import type { EmailVerificationCode } from "../../../../domain/entities";
 
 export interface IEmailVerificationCodeRepository {
 	findByUserId(userId: EmailVerificationCode["userId"]): Promise<EmailVerificationCode | null>;
