+++ apps/api/src/interface-adapter/repositories/oauth-account/interfaces/oauth-account.repository.interface.ts
@@ -1,4 +1,4 @@
-import type { OAuthAccount } from "../../../../domain/entities/oauth-account";
+import type { OAuthAccount } from "../../../../domain/entities";
 
 export interface IOAuthAccountRepository {
 	findByUserId(userId: OAuthAccount["userId"]): Promise<OAuthAccount | null>;
