+++ apps/api/src/interface-adapter/repositories/oauth-account/tests/update-by-user-id.test.ts
@@ -1,6 +1,6 @@
 import { env } from "cloudflare:test";
 import { beforeAll, describe, expect, test } from "vitest";
-import { OAuthAccount } from "../../../../domain/entities/oauth-account";
+import { OAuthAccount } from "../../../../domain/entities";
 import { DrizzleService } from "../../../../infrastructure/drizzle";
 import { type DatabaseOAuthAccount, OAuthAccountTableHelper, UserTableHelper } from "../../../../tests/helpers";
 import { OAuthAccountRepository } from "../oauth-account.repository";
