+++ apps/api/src/interface-adapter/repositories/user-credential/interfaces/user-credential.repository.interface.ts
@@ -1,4 +1,4 @@
-import type { UserCredential } from "../../../../domain/entities/user-credential";
+import type { UserCredential } from "../../../../domain/entities";
 
 export interface IUserCredentialRepository {
 	create(
