+++ apps/api/src/interface-adapter/repositories/user/tests/update.test.ts
@@ -1,6 +1,6 @@
 import { env } from "cloudflare:test";
 import { beforeAll, describe, expect, test } from "vitest";
-import { User } from "../../../../domain/entities/user";
+import { User } from "../../../../domain/entities";
 import { DrizzleService } from "../../../../infrastructure/drizzle";
 import { type DatabaseUser, UserTableHelper } from "../../../../tests/helpers";
 import { UserRepository } from "../user.repository";
