+++ apps/api/src/modules/auth-guard/auth-guard.plugin.ts
@@ -3,8 +3,7 @@ import { ValidateSessionUseCase } from "../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../common/constants";
 import { readBearerToken } from "../../common/utils";
 import { isErr } from "../../common/utils";
-import type { Session } from "../../domain/entities/session";
-import type { User } from "../../domain/entities/user";
+import type { Session, User } from "../../domain/entities";
 import { DrizzleService } from "../../infrastructure/drizzle";
 import { SessionRepository } from "../../interface-adapter/repositories/session";
 import { UserRepository } from "../../interface-adapter/repositories/user";
